Dev[Data]: Исправил смысл сущности
Убрал из сущности преобразования из dto, вынес эти преобразования в маппер. Сделал главный виджет Stateful для dispose контроллера скролла

Related to: Alan Turing
