+++ lib/core/data/repositories/to_do_list_repository_impl.dart
@@ -2,9 +2,9 @@ import 'dart:io' show Platform;
 import 'package:dartz/dartz.dart';
 import 'package:demetiapp/core/data/datasources/local/task_local_datasource.dart';
 import 'package:demetiapp/core/data/datasources/remote/task_remote_datasource.dart';
-import 'package:demetiapp/core/data/models/task_api_model.dart';
-import 'package:demetiapp/core/data/models/task_local_model.dart';
-import 'package:demetiapp/core/data/models/task_mapper.dart';
+import 'package:demetiapp/core/data/dto/task_api_model.dart';
+import 'package:demetiapp/core/data/dto/task_local_model.dart';
+import 'package:demetiapp/core/data/dto/task_mapper.dart';
 import 'package:demetiapp/core/domain/entities/task_entity.dart';
 import 'package:demetiapp/core/domain/repository/todo_list_repository.dart';
 import 'package:demetiapp/core/error/exception.dart';
